+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Fixed an internal bug that caused `Rebuilder` to block cascading updates
+- Made `read` and `maybeRead` not depend on the `InheritedProvider`, added new methods `depend` and `maybeDepend` to genuinely depend on a provider
+
 ## 0.0.1
 
 First release.